requirements.txt
# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
statsmodels>=0.14.0

# Web Framework
streamlit>=1.28.0

# HTTP & APIs
httpx>=0.25.0

# Configuration & Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0.0

# Data Storage
pyarrow>=13.0.0
sqlalchemy>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Correlation & Copulas
copulas>=0.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Development
black>=23.0.0
ruff>=0.1.0
mypy>=1.5.0
